# apps/api/openapi.yml
openapi: 3.1.0
info:
  title: Board Game Rules Assistant API
  version: 1.0.0
  description: API for rulebook upload, indexing, and rules Q&A.
servers:
  - url: http://127.0.0.1:8000
    description: Local development
tags:
  - name: Chats
    description: Chat conversation lifecycle
  - name: Health
    description: API health checks
  - name: Ingestion
    description: Rulebook upload and indexing
  - name: Retrieval
    description: Similarity search over indexed rulebook chunks
paths:
  /chats:
    get:
      tags:
        - Chats
      summary: List chat summaries
      description: Returns conversation ids and titles ordered from newest to oldest.
      responses:
        "200":
          description: Chat summaries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetChatsResponse"
        "500":
          description: Unexpected chat listing failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Chats
      summary: Create an empty conversation
      description: Creates a conversation without requiring a request body.
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateChatResponse"
        "500":
          description: Unexpected conversation creation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chats/{id}:
    get:
      tags:
        - Chats
      summary: Get a chat history
      description: Returns one conversation and its persisted messages ordered from oldest to newest.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetChatResponse"
        "404":
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value:
                    error: Chat not found
        "500":
          description: Unexpected chat history failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Chats
      summary: Hard delete a chat
      description: Permanently removes the conversation and all of its messages.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Chat deleted
        "404":
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value:
                    error: Chat not found
        "500":
          description: Unexpected chat deletion failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Returns a basic status response for the API.
      responses:
        "200":
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /ready:
    get:
      tags:
        - Health
      summary: Readiness check
      description: Returns success only when required persistence dependencies are available.
      responses:
        "200":
          description: API dependencies are ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "503":
          description: A required dependency is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rulebooks:
    post:
      tags:
        - Ingestion
      summary: Create a rulebook from an uploaded PDF
      description: Uploads and indexes a PDF rulebook. With PostgreSQL persistence, a successful request also stores the original PDF and its metadata for subsequent listing.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadRulebookRequest"
      responses:
        "200":
          description: Rulebook indexed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadRulebookResponse"
        "400":
          description: Invalid upload request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missingFile:
                  value:
                    error: file is required
                invalidFileType:
                  value:
                    error: Only PDF files are allowed
                invalidSplitterParams:
                  value:
                    error: chunkOverlap must be less than chunkSize
        "413":
          description: File exceeds upload size limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                fileTooLarge:
                  value:
                    error: File exceeds the upload size limit
        "500":
          description: Unexpected ingestion failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags:
        - Ingestion
      summary: List indexed rulebooks
      description: Returns rulebook metadata from the selected repository, newest first. PDF content is not loaded or returned.
      responses:
        "200":
          description: Indexed rulebooks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListRulebooksResponse"
        "500":
          description: Unexpected list failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rulebooks/{id}:
    delete:
      tags:
        - Ingestion
      summary: Delete a persisted rulebook
      description: Removes the indexed vectors, rulebook metadata, and stored PDF.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Rulebook deleted
        "404":
          description: Rulebook not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value:
                    error: Rulebook not found
        "500":
          description: Unexpected delete failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /retrieval/search:
    post:
      tags:
        - Retrieval
      summary: Search indexed rulebook chunks
      description: Runs vector similarity search for an existing conversation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetrievalSearchRequest"
      responses:
        "200":
          description: Matching rulebook chunks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetrievalSearchResponse"
        "400":
          description: Invalid search request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missingQuery:
                  value:
                    error: query is required
        "404":
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value:
                    error: Conversation not found
        "500":
          description: Unexpected retrieval failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    CreateChatResponse:
      type: object
      required:
        - conversationId
      properties:
        conversationId:
          type: string
          format: uuid
          description: Server-generated conversation identifier.
      additionalProperties: false
    ChatSummary:
      type: object
      required:
        - conversationId
        - title
      properties:
        conversationId:
          type: string
          format: uuid
          description: Server-generated conversation identifier.
        title:
          type: string
          minLength: 1
          description: Conversation title.
      additionalProperties: false
    GetChatsResponse:
      type: object
      required:
        - chats
      properties:
        chats:
          type: array
          items:
            $ref: "#/components/schemas/ChatSummary"
      additionalProperties: false
    ChatMessage:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum:
            - user
            - assistant
        content:
          type: string
      additionalProperties: false
    GetChatResponse:
      type: object
      required:
        - conversationId
        - title
        - messages
      properties:
        conversationId:
          type: string
          format: uuid
          description: Server-generated conversation identifier.
        title:
          type: string
          minLength: 1
          description: Conversation title.
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
      additionalProperties: false
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message.
        stack:
          type: string
          description: Development-only stack trace for unhandled errors.
      additionalProperties: false
    HealthResponse:
      type: object
      required:
        - status
        - service
      properties:
        status:
          type: string
          const: ok
        service:
          type: string
          const: board-game-rules-assistant-api
      additionalProperties: false
    UploadRulebookRequest:
      type: object
      required:
        - gameName
        - file
      properties:
        gameName:
          type: string
          minLength: 1
          description: Board game name associated with the uploaded PDF.
          example: Catan
        file:
          type: string
          format: binary
          description: PDF rulebook file. The API accepts only `.pdf` files with `application/pdf` MIME type.
        chunkSize:
          type: integer
          minimum: 1
          description: Optional override for text splitter chunk size.
          example: 1000
        chunkOverlap:
          type: integer
          minimum: 0
          description: Optional override for text splitter chunk overlap. Must be less than `chunkSize` when both are provided.
          example: 150
      additionalProperties: false
    RulebookSummary:
      type: object
      required:
        - id
        - gameName
        - pdfName
        - fileSize
      properties:
        id:
          type: string
          format: uuid
          description: Server-generated rulebook identifier.
        gameName:
          type: string
          minLength: 1
          description: Board game name provided at upload time.
          example: Catan
        pdfName:
          type: string
          minLength: 1
          description: Original PDF filename.
          example: catan-rulebook.pdf
        fileSize:
          type: integer
          minimum: 0
          description: Uploaded file size in bytes.
          example: 1840000
      additionalProperties: false
    UploadRulebookResponse:
      type: object
      required:
        - id
        - gameName
        - pdfName
        - fileSize
        - status
        - documentCount
        - chunkCount
      properties:
        id:
          type: string
          format: uuid
          description: Server-generated rulebook identifier.
        gameName:
          type: string
          minLength: 1
          description: Board game name provided at upload time.
          example: Catan
        pdfName:
          type: string
          minLength: 1
          description: Original PDF filename.
          example: catan-rulebook.pdf
        fileSize:
          type: integer
          minimum: 0
          description: Uploaded file size in bytes.
          example: 1840000
        status:
          type: string
          const: completed
          description: Ingestion status for this synchronous upload request.
        documentCount:
          type: integer
          minimum: 0
          description: Number of page-level documents extracted from the PDF.
        chunkCount:
          type: integer
          minimum: 0
          description: Number of chunks indexed into the in-memory vector store.
      additionalProperties: false
    ListRulebooksResponse:
      type: object
      required:
        - rulebooks
      properties:
        rulebooks:
          type: array
          description: Rulebooks indexed during the current API process lifetime.
          items:
            $ref: "#/components/schemas/RulebookSummary"
      additionalProperties: false
    RetrievalSearchRequest:
      type: object
      required:
        - conversationId
        - query
      properties:
        conversationId:
          type: string
          format: uuid
          description: Server-generated identifier returned by POST /chats.
          example: 11111111-1111-4111-8111-111111111111
        query:
          type: string
          minLength: 1
          description: Natural-language search query.
          example: How many resources does a city produce?
      additionalProperties: false
    RetrievalMatchMetadata:
      type: object
      properties:
        documentId:
          type: string
          format: uuid
          description: Rulebook id attached during upload.
        pageNumber:
          type: integer
          minimum: 1
          description: Source PDF page number when available.
        source:
          type: string
          description: Source PDF filename or public web URL when available.
          example: catan-rulebook.pdf
      additionalProperties: false
    RetrievalMatch:
      type: object
      required:
        - origin
        - content
        - metadata
      properties:
        origin:
          type: string
          enum:
            - rulebook
            - public_web
          description:
            Whether the match came from an indexed rulebook or public web
            search.
        content:
          type: string
          description: Retrieved chunk text.
        metadata:
          $ref: "#/components/schemas/RetrievalMatchMetadata"
      additionalProperties: false
    RetrievalSearchResponse:
      type: object
      required:
        - title
        - answer
        - matches
      properties:
        title:
          type: string
          minLength: 1
          description: Persisted conversation title.
        answer:
          type: string
          description: Agent-generated answer grounded in retrieved rulebook context.
        matches:
          type: array
          items:
            $ref: "#/components/schemas/RetrievalMatch"
      additionalProperties: false